remove minify, reset

+++ build.js
@@ -27,8 +27,7 @@ const dist = path.join(__dirname, "dist");
 const ext  = path.join(__dirname, "extension.zip");
 
 /* clear dist */ {
-    if(fs.existsSync(dist))
-        fs.rmSync(dist, {recursive: true});
+    fs.existsSync(dist) && fs.rmSync(dist, {recursive: true});
     fs.mkdirSync(dist);
 
     !fs.existsSync(ext) || fs.rmSync(ext, {recursive: true});
@@ -37,23 +36,6 @@ const ext  = path.join(__dirname, "extension.zip");
 /* copy src to zip */ {
     for(const file of fs.readdirSync(src))
         fs.copyFileSync(path.join(src, file), path.join(dist, file));
-}
-
-/* minify */ {
-    for(const file of ["options.css", "options.js", "index.js", "style.css"])
-        fs.writeFileSync(path.join(dist, file), fs.readFileSync(path.join(dist, file), "utf-8")
-            .replace(/(?<!^)\/\*.*\*\//g, '') // /* comments (except first copyright)
-            .replace(/ \/\/.*$/gm,'') // // comments
-            .replace(/ +/gm, ' ') // extra spaces
-            .replace(/^ +/gm, '') // leading space
-            .replace(/\r?\n/gm, '') // new line
-            .trim());
-    for(const file of ["options.html"])
-        fs.writeFileSync(path.join(dist, file), fs.readFileSync(path.join(dist, file), "utf-8")
-            .replace(/ +/gm, ' ') // extra spaces
-            .replace(/^ +/gm, '') // leading space
-            .replace(/\r?\n/gm, '') // new line
-            .trim());
 }
 
 const zip = new AdmZip();
